requirements.txt
asgiref==3.11.0
Django>=5.2,<6.0
djangorestframework==3.16.1
gunicorn==23.0.0
Markdown==3.10
packaging==25.0
sqlparse==0.5.5
